comments added

+++ src/comments/comments.service.ts
@@ -4,17 +4,33 @@ import { CreateCommentDto } from './dto/create-comment.dto';
 import { UpdateCommentDto } from './dto/update-comment.dto';
 import { Comment } from './entities/comment.entity';
 import { InjectRepository } from '@nestjs/typeorm';
+import { Book } from 'src/books/entities/book.entity';
 
 @Injectable()
 export class CommentsService {
   constructor(
     @InjectRepository(Comment)
     private readonly commentRepo: Repository<Comment>,
+    @InjectRepository(Book) private readonly bookRepo: Repository<Book>,
   ) {}
 
   async create(createCommentDto: CreateCommentDto): Promise<Comment> {
     const comment = this.commentRepo.create(createCommentDto);
-    return await this.commentRepo.save(comment);
+    const createdComment = await this.commentRepo.save(comment);
+
+    const book = await this.bookRepo.findOne({
+      relations: ['ratings'],
+      where: { id: createCommentDto.bookId },
+    });
+
+    if (!book) {
+      throw new NotFoundException(`Book not found!`);
+    }
+
+    book.comments.push(createdComment);
+    await this.bookRepo.save(book);
+
+    return createdComment;
   }
 
   async findOne(id: string): Promise<Comment | null> {
@@ -38,6 +54,26 @@ export class CommentsService {
       comment.text = updateCommentDto.text;
     }
 
+    const book = await this.bookRepo.findOne({
+      relations: ['comments'],
+      where: { id: comment.book },
+    });
+
+    if (!book) {
+      throw new NotFoundException(`Book not found!`);
+    }
+
+    const commentToUpdateIndex = book.comments.findIndex(
+      (c) => c.id === comment.id,
+    );
+
+    if (commentToUpdateIndex === -1) {
+      throw new NotFoundException(`Comment not found in book's comments!`);
+    }
+
+    book.comments[commentToUpdateIndex] = comment;
+    await this.bookRepo.save(book);
+
     return await this.commentRepo.save(comment);
   }
 
@@ -46,10 +82,20 @@ export class CommentsService {
     if (!comment) {
       throw new NotFoundException(`Comment not found!`);
     }
-    return await this.commentRepo.remove(comment);
-  }
 
-  async findAllByBook(bookId: string): Promise<Comment[]> {
-    return await this.commentRepo.find({ where: { book: bookId } });
+    const book = await this.bookRepo.findOne({
+      relations: ['comments'],
+      where: { id: comment.book },
+    });
+
+    if (!book) {
+      throw new NotFoundException(`Book not found!`);
+    }
+
+    book.comments = book.comments.filter((c) => c.id !== id);
+
+    await this.bookRepo.save(book);
+
+    return await this.commentRepo.remove(comment);
   }
 }
